Crypt

Description

The salsa20 encryption and decryption.

Documentation

cryptBlockV1 :: [Word32] -> [Word32] -> [Word32] -> Word8 -> [Word32]

Encrypt or decrypt a message with a single 16 bytes key resulting in a list of the same length.

cryptBlockV2 :: [Word32] -> [Word32] -> [Word32] -> [Word32] -> Word8 -> [Word32]

Encrypt or decrypt a message with two 16 bytes key resulting in a list of the same length.

iOver64Display :: String -> [String]

Display the calculation of an index over 64. See iOver64.

nonceAndiOver64Display :: [String] -> String -> [String]

Join the nonce and the calculated iOver64Display.

cryptV1Display :: [String] -> [String] -> String -> [String]

Given a single 16 bytes key, a nonce and an index get the salsa20 expanded matrix of it.

cryptV2Display :: [String] -> [String] -> [String] -> String -> [String]

Given two 16 bytes keys, a nonce and an index get the salsa20 expanded matrix of it.

cryptBlockV1Display :: [String] -> [String] -> [String] -> Int -> [String]

Encrypt or decrypt a message with a single 16 bytes key resulting in a list of the same length.

cryptBlockV1Equations :: [String] -> [String] -> [String] -> Int -> [String]

Display the output of cryptBlockV1 as string equations.