Hash

Description

The salsa20 core function and what we call the hash expressions.

The hash expression is the extended core expression. It is formed by reducing the input from 64 to 16 using littleendian functions and then aument the result back to 64.

Documentation

coreCompute :: [Word32] -> [Word32]

The core expression computed.

coreDisplay :: [String] -> [String]

The core expression as a string.

core1Compute :: [Word32] -> [Word32]

The core1 expression computed.

core1Display :: [String] -> [String]

The core1 expression as a string.

core1Equations :: [String] -> [String]

The core1 expression as a list of equations.

core2Compute :: [Word32] -> [Word32]

The core2 expression computed.

core2Display :: [String] -> [String]

The core2 expression as a string.

core2Equations :: [String] -> [String]

The core2 expression as a list of equations.

salsa20Compute :: [Word32] -> [Word32]

The salsa20 expression computed.

salsa20Display :: [String] -> [String]

The salsa20 expression as a string using core1Display which is only one round of doubleround.

salsa20Equations :: [String] -> [String]

The salsa20 expression as a list of equations.

salsa20powerCompute :: [Word32] -> Int -> [Word32]

Execute salsa20 a specified number of times, this is not part of the protocol and just used in a test case.