LittleEndian

Description

Documentation

littleendian :: [Word32] -> Word32

Encode a vector as a word using little-endian byte order.

extractBytes4 :: Word32 -> [Word32]

Extract 4 bytes from a Word32.

This function is implemented as simple haskell (no syntactic sugar) because of liquidhaskell refinements in list lengths.

extractBytes8 :: Word32 -> [Word32]

Extract 8 bytes from a Word32.

This function is implemented as simple haskell (no syntactic sugar) because of liquidhaskell refinements in list lengths.

displayBytes4 :: String -> [String]

Display a 4 bytes from a string as a list of strings.

This function is implemented as simple haskell (no syntactic sugar) because of liquidhaskell refinements in list lengths.

displayBytes8 :: String -> [String]

Display a 8 bytes from a string as a list of strings.

This function is implemented as simple haskell (no syntactic sugar) because of liquidhaskell refinements in list lengths.

reduce :: [Word32] -> [Word32]

Reduce a matrix of 64 elements to a matrix of 16 elements by using littleendian encoding.

reduceDisplay :: [String] -> [String]

Reduce a matrix of 64 elements to a matrix of 16 elements by using littleendianDisplay encoding.

reduceKeelung :: [UInt 32] -> [UInt 32]

Reduce a matrix of 64 elements to a matrix of 16 elements by using littleendianKeelung encoding.

aument :: [Word32] -> [Word32]

Aument a matrix of 16 elements to one of 64 elements by using extractBytes.

aumentDisplay :: [String] -> [String]

Aument a matrix of 16 elements to one of 64 elements by using displayBytes.

aumentKeelung :: [UInt 32] -> [UInt 32]

Aument a matrix of 16 elements to one of 64 elements by using extractBytesKeelung.