Expansion

Description

There are two forms of expansion supported by Salsa20 given by key size. The 16 bytes (128 bits) key version is called expand16 and the 32 bytes (256 bits) key version is called expand32. The 32 bytes version seems to be the most used one.

Documentation

expand32Compute :: [Word32] -> [Word32] -> [Word32] -> [Word32]

The expansion function where we have two 16 bytes k's (k0 and k1).

expand32Display :: [String] -> [String] -> [String] -> [String]

The expansion function displayed where we have two 16 bytes (k0 and k1).

expand32Equations :: [String] -> [String] -> [String] -> [String]

The expansion function equations where we have two 16 bytes (k0 and k1).

expand16Compute :: [Word32] -> [Word32] -> [Word32]

The expansion function computed where we have one 16 bytes (k).

expand16Display :: [String] -> [String] -> [String]

The expansion function displayed where we have one 16 bytes (k).

expand16Equations :: [String] -> [String] -> [String]

The expansion function equations where we have one 16 bytes (k).