Utils

Description

General utility functions used to create the salsa20 cipher.

Documentation

littleendian :: [Word32] -> Word32

Encode a vector as a word using protocol specified littleendian.

littleendianInv :: Word32 -> [Word32]

The inverse of littleendian. Implemented as specified in https://crypto.stackexchange.com/a/22314.

reduce :: [Word32] -> [Word32]

aument :: [Word32] -> [Word32]

modMatrix :: [Word32] -> [Word32] -> [Word32]

Given two matrices, do modulo addition on each of the elements.

littleendianInv4Crypt :: Word32 -> [Word32]

The littleendian function used in encryption/decryption where a number is obtained given a list of 8 bytes.