Utils

Description

General utility functions used to create the salsa20 cipher.

Documentation

littleendian :: [Word32] -> Word32

Encode a vector as a word using protocol specified littleendian.

littleendianInv :: Word32 -> [Word32]

The inverse of littleendian. Implemented as specified in https://crypto.stackexchange.com/a/22314.

reduce :: [Word32] -> [Word32]

Reduce a matrix of 64 elements to a matrix of 16 elements by using littleendian encoding.

reduceDisplay :: [String] -> [String]

Reduce a matrix of 64 elements to a matrix of 16 elements by using littleendianDisplay encoding.

aument :: [Word32] -> [Word32]

Aument a matrix of 16 elements to one of 64 elements by using littleendianInv.

aumentDisplay :: [String] -> [String]

Aument a matrix of 16 elements to one of 64 elements by using littleendianInvDisplay.

modMatrix :: [Word32] -> [Word32] -> [Word32]

Given two matrices, do modulo addition on each of the elements.

littleendianInv4Crypt :: Word32 -> [Word32]

The littleendian function used in encryption/decryption where a list of 8 bytes is obtained from a number.

littleendianInv4CryptDisplay :: String -> [String]

The littleendianInv4Crypt as a string.

stringListToNumberList :: [String] -> [Word32]

Convert a list of strings to a list of numbers if possible.

numberListToStringList :: [Word32] -> [String]

Convert a list of numbers to a list of strings. This is always possible.

transpose :: [a] -> [a]

Transpose a 4x4 matrix type.

index0 :: Int

The number 0 as an integer. Used as list index in certain cases.

modMatrixDisplay :: [String] -> [String] -> [String]

Given two matrices, do modulo addition on each of the elements.

numberListToEitherList :: [Word32] -> [Either Word32 String]

Convert a list of numbers to an list of Either type.

stringListToEitherList :: [String] -> [Either Word32 String]

Convert a list of strings to a list of Either type.

eitherListToNumberList :: [Either Word32 String] -> [Word32]

Convert a list of Either type to an list of numbers.

eitherListToStringList :: [Either Word32 String] -> [String]

Convert a list of Either type to a list of strings.

listOfNumbersToListOfHex :: [Word32] -> [String]

(>>>) :: Word32 -> Int -> Word32

The shiftR operator.

(&) :: Word32 -> Word32 -> Word32

The bitwise and operator.